quick fix

+++ tableCreation.sql
@@ -1,9 +1,10 @@
-DROP DATABASE Cinema
+IF EXISTS(select * from sys.databases where name='Cinema')
+	DROP DATABASE Cinema
 CREATE DATABASE Cinema
-GO
-USE Cinema 
 
-DROP TABLE Movies
+IF OBJECT_ID(N'dbo.Movies', N'U') IS NOT NULL  
+   DROP TABLE [dbo].[Movies];  
+GO
 CREATE TABLE [dbo].[Movies]
 (
   [movieID] INT  NOT NULL IDENTITY(1,1) PRIMARY KEY,
@@ -19,7 +20,7 @@ CREATE TABLE [dbo].[Movies]
 INSERT INTO Movies (movieTitle,genre,releaseYear,director,length,country,onDisplay)VALUES
 ('Godfather','crime',1972,'Francis Ford Coppola',170,'USA','YES'),
 (' The Shawshank Redemption','drama',1994,'Frank Darabont',142,'USA','YES'),
-([Schindler's List],'history',1993,'Steven Spielberg',195,'USA','YES'),
+('Schindlers List','history',1993,'Steven Spielberg',195,'USA','YES'),
 ('Forrest Gump','comedy',1994,'Robert Zemeckis',142,'USA','YES'),
 ('12 Angry Men','crime',1957,'Sidney Lumet',96,'USA','NO'),
 ('Star Wars: Episode IV - A New Hope','Fantasy',1977,'George Lucas',121,'USA','YES'),
@@ -34,4 +35,4 @@ INSERT INTO Movies (movieTitle,genre,releaseYear,director,length,country,onDispl
 ('Léon','thriller',1994,'Luc Besson',110,'France','NO'),
 ('The Good The Bad and The Ugly','western',1966,'Todd Phillips',179,'Italy','NO'),
 ('Inglourious Basterds','war',2009,'Quentin Tarantino',170,'USA','YES'),
-('Avatar: The Way of Water','sci-fi',2022,'James Cameron',192,'USA','YES')
+('Avatar: The Way of Water','sci-fi',2022,'James Cameron',192,'USA','YES')
